+++ CHANGELOG.md
@@ -37,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#1063](https://github.com/spegel-org/spegel/pull/1063) Replace transport with option to set OCI client.
 - [#1065](https://github.com/spegel-org/spegel/pull/1065) Move web to pkg and remove metrics dependency.
 - [#1066](https://github.com/spegel-org/spegel/pull/1066) Wait for router to be ready before tracking state.
+- [#1067](https://github.com/spegel-org/spegel/pull/1067) Keep files in registry directory when prepending to file.
   
 ### Deprecated